+++ lib/src/platform_info.dart
@@ -1,0 +1,6 @@
+import 'platform_info_stub.dart' if (dart.library.io) 'platform_info_io.dart'
+    as implementation;
+
+String get operatingSystem => implementation.operatingSystem;
+
+String get operatingSystemVersion => implementation.operatingSystemVersion;