+++ lib/src/platform_info_io.dart
@@ -1,0 +1,5 @@
+import 'dart:io';
+
+String get operatingSystem => Platform.operatingSystem;
+
+String get operatingSystemVersion => Platform.operatingSystemVersion;